Moved FullTextSessionFactory to org.openmrs.api.db package
