+++ api/src/main/java/org/openmrs/api/db/FullTextSessionFactory.java
@@ -7,7 +7,7 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs;
+package org.openmrs.api.db;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
