+++ api/src/main/java/org/openmrs/api/db/hibernate/search/LuceneQuery.java
@@ -28,7 +28,7 @@ import org.hibernate.Session;
 import org.hibernate.search.FullTextQuery;
 import org.hibernate.search.FullTextSession;
 import org.hibernate.search.query.dsl.QueryBuilder;
-import org.openmrs.FullTextSessionFactory;
+import org.openmrs.api.db.FullTextSessionFactory;
 import org.openmrs.PatientIdentifier;
 import org.openmrs.PersonAttribute;
 import org.openmrs.PersonName;
